clean notification controller and voter #2823

+++ src/Controller/Back/NotificationController.php
@@ -34,18 +34,19 @@ class NotificationController extends AbstractController
     public function read(
         Request $request,
         EntityManagerInterface $entityManager,
-        NotificationRepository $notificationRepository,
     ) {
         /** @var User $user */
         $user = $this->getUser();
         if ($request->get('selected_notifications')) {
             if ($this->isCsrfTokenValid('mark_as_read_'.$user->getId(), $request->get('csrf_token'))) {
-                $this->markSelectedAsRead($entityManager, $notificationRepository, explode(',', $request->get('selected_notifications')));
+                $q = $entityManager->createQuery('UPDATE '.Notification::class.' n SET n.isSeen = true WHERE n.user = :user AND n.id IN (:ids)');
+                $q->execute(['user' => $user, 'ids' => explode(',', $request->get('selected_notifications'))]);
                 $this->addFlash('success', 'Les notifications sélectionnées ont été marquées comme lues.');
             }
         } else {
             if ($this->isCsrfTokenValid('mark_as_read_'.$user->getId(), $request->get('mark_as_read'))) {
-                $this->markAllAsRead($entityManager);
+                $q = $entityManager->createQuery('UPDATE '.Notification::class.' n SET n.isSeen = true WHERE n.user = :user');
+                $q->execute(['user' => $user]);
                 $this->addFlash('success', 'Toutes les notifications ont été marquées comme lues.');
             }
         }
@@ -57,13 +58,13 @@ class NotificationController extends AbstractController
     public function delete(
         Request $request,
         EntityManagerInterface $entityManager,
-        NotificationRepository $notificationRepository,
     ) {
         /** @var User $user */
         $user = $this->getUser();
         if ($request->get('selected_notifications')) {
             if ($this->isCsrfTokenValid('delete_notifications_'.$user->getId(), $request->get('csrf_token'))) {
-                $this->deleteSelected($entityManager, $notificationRepository, explode(',', $request->get('selected_notifications')));
+                $q = $entityManager->createQuery('DELETE FROM '.Notification::class.' n WHERE n.user = :user AND n.id IN (:ids)');
+                $q->execute(['user' => $user, 'ids' => explode(',', $request->get('selected_notifications'))]);
                 $this->addFlash('success', 'Les notifications sélectionnées ont été supprimées.');
             }
         } else {
@@ -71,64 +72,13 @@ class NotificationController extends AbstractController
                 'delete_all_notifications_'.$user->getId(),
                 $request->get('delete_all_notifications')
             )) {
-                $this->deleteAllNotifications($entityManager);
+                $q = $entityManager->createQuery('DELETE FROM '.Notification::class.' n WHERE n.user = :user');
+                $q->execute(['user' => $user]);
                 $this->addFlash('success', 'Toutes les notifications ont été supprimées.');
             }
         }
 
         return $this->redirectToRoute('back_notifications_list');
-    }
-
-    private function markSelectedAsRead(
-        EntityManagerInterface $em,
-        NotificationRepository $notificationRepository,
-        array $listIdNotifications
-    ) {
-        $notifications = $notificationRepository->findUserNotificationsInList($this->getUser(), $listIdNotifications);
-        foreach ($notifications as $notification) {
-            $this->denyAccessUnlessGranted('NOTIF_MARK_AS_READ', $notification);
-            $notification->setIsSeen(true);
-            $em->persist($notification);
-        }
-        $em->flush();
-    }
-
-    private function markAllAsRead($em)
-    {
-        /** @var User $user */
-        $user = $this->getUser();
-        $notifications = $user->getNotifications();
-        $notifications->filter(function (Notification $notification) use ($em) {
-            $this->denyAccessUnlessGranted('NOTIF_MARK_AS_READ', $notification);
-            $notification->setIsSeen(true);
-            $em->persist($notification);
-        });
-        $em->flush();
-    }
-
-    private function deleteSelected(
-        EntityManagerInterface $em,
-        NotificationRepository $notificationRepository,
-        array $listIdNotifications
-    ) {
-        $notifications = $notificationRepository->findUserNotificationsInList($this->getUser(), $listIdNotifications);
-        foreach ($notifications as $notification) {
-            $this->denyAccessUnlessGranted('NOTIF_DELETE', $notification);
-            $em->remove($notification);
-        }
-        $em->flush();
-    }
-
-    private function deleteAllNotifications($em)
-    {
-        /** @var User $user */
-        $user = $this->getUser();
-        $notifications = $user->getNotifications();
-        $notifications->filter(function (Notification $notification) use ($em) {
-            $this->denyAccessUnlessGranted('NOTIF_DELETE', $notification);
-            $em->remove($notification);
-        });
-        $em->flush();
     }
 
     #[Route('/notifications/{id}/supprimer', name: 'back_notifications_delete_notification')]
@@ -141,9 +91,9 @@ class NotificationController extends AbstractController
         if (!$notification) {
             return $this->redirectToRoute('back_notifications_list');
         }
-
-        $this->denyAccessUnlessGranted('NOTIF_DELETE', $notification);
-        if ($this->isCsrfTokenValid('back_delete_notification_'.$notification->getId(), $request->get('_token'))) {
+        /** @var User $user */
+        $user = $this->getUser();
+        if ($notification->getUser()->getId() === $user->getId() && $this->isCsrfTokenValid('back_delete_notification_'.$notification->getId(), $request->get('_token'))) {
             $em->remove($notification);
             $em->flush();
             $this->addFlash('success', 'Notification supprimée avec succès');
